+++ changelog/changelog.md
@@ -1,8 +1,15 @@
 ## [v1.0.0] - 2024-12-02
 first docker release
 
-## [v1.0.0] - 2025-3-23
+## [v1.1.0] - 2025-3-23
 added erosion for those with large final value
 
-## [v1.0.0] - 2025-3-24
+## [v1.2.0] - 2025-3-24
 added sobel filter option
+
+## [v1.3.0] - 2025-3-25
+added segmentation points and metrics display in interactive plot
+
+
+## [v1.3.0] - 2025-3-26
+added skeleton length and spline length